v2.0.0 Create Locations module
Create the migration and model of Locations module, also create the Table, Form and Delete component of this module, also create his data-card and form-cad blade components. Create the permissions of it and his web route as 'locations', also add the option button in sys-navbar, in this nav add the 'wire:navigate' feature. Besides, update the CommonUtils class and the messages.php languages files to add new keys. Important, don't code the functionality of Table, Form and Delete of Location module, just create and enable it - 20-07-2024